requirements.txt
fastapi==0.115.2
fastapi[standard]==0.115.2
elevenlabs==1.10.0

# env, tests
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2  # test async
pytest-asyncio==0.24.0

# code
black==24.10.0
pylint==3.3.1

# crypt
argon2-cffi==23.1.0
passlib==1.7.4
PyJWT==2.9.0
